// packages/tddy-web/src/gen/terminal_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts"
// @generated from file terminal.proto (package terminal, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file terminal.proto.
 */
export const file_terminal: GenFile = /*@__PURE__*/
  fileDesc("Cg50ZXJtaW5hbC5wcm90bxIIdGVybWluYWwiHQoNVGVybWluYWxJbnB1dBIMCgRkYXRhGAEgASgMIh4KDlRlcm1pbmFsT3V0cHV0EgwKBGRhdGEYASABKAwyXAoPVGVybWluYWxTZXJ2aWNlEkkKEFN0cmVhbVRlcm1pbmFsSU8SFy50ZXJtaW5hbC5UZXJtaW5hbElucHV0GhgudGVybWluYWwuVGVybWluYWxPdXRwdXQoATABYgZwcm90bzM");

/**
 * @generated from message terminal.TerminalInput
 */
export type TerminalInput = Message<"terminal.TerminalInput"> & {
  /**
   * @generated from field: bytes data = 1;
   */
  data: Uint8Array;
};

/**
 * Describes the message terminal.TerminalInput.
 * Use `create(TerminalInputSchema)` to create a new message.
 */
export const TerminalInputSchema: GenMessage<TerminalInput> = /*@__PURE__*/
  messageDesc(file_terminal, 0);

/**
 * @generated from message terminal.TerminalOutput
 */
export type TerminalOutput = Message<"terminal.TerminalOutput"> & {
  /**
   * @generated from field: bytes data = 1;
   */
  data: Uint8Array;
};

/**
 * Describes the message terminal.TerminalOutput.
 * Use `create(TerminalOutputSchema)` to create a new message.
 */
export const TerminalOutputSchema: GenMessage<TerminalOutput> = /*@__PURE__*/
  messageDesc(file_terminal, 1);

/**
 * @generated from service terminal.TerminalService
 */
export const TerminalService: GenService<{
  /**
   * @generated from rpc terminal.TerminalService.StreamTerminalIO
   */
  streamTerminalIO: {
    methodKind: "bidi_streaming";
    input: typeof TerminalInputSchema;
    output: typeof TerminalOutputSchema;
  },
}> = /*@__PURE__*/
  serviceDesc(file_terminal, 0);